+++ src/Data/Morpheus/Types/Internal/Resolver.hs
@@ -17,9 +17,7 @@ module Data.Morpheus.Types.Internal.Resolver
   , Event(..)
   , GQLRootResolver(..)
   , UnSubResolver
-  , GQLFail(..)
   , ResponseT
-  , failResolveT
   , Resolver(..)
   , ResolvingStrategy(..)
   , MapGraphQLT(..)
@@ -28,18 +26,18 @@ module Data.Morpheus.Types.Internal.Resolver
   , toResponseRes
   , withObject
   , Resolving(..)
+  , liftM
+  , liftEitherM
   ) where
 
 import           Control.Monad.Trans.Except                 (ExceptT (..), runExceptT, withExceptT)
 import           Data.Maybe                                 (fromMaybe)
 import           Data.Semigroup                             ((<>))
-import           Data.Text                                  (pack, unpack)
 
 -- MORPHEUS
 import           Data.Morpheus.Error.Selection              (resolverError, subfieldsNotSelected)
 import           Data.Morpheus.Types.Internal.AST.Selection (Selection (..), SelectionRec (..), SelectionSet,
                                                              ValidSelection)
-import           Data.Morpheus.Types.Internal.Base          (Message)
 import           Data.Morpheus.Types.Internal.Data          (Key, MUTATION, OperationType, QUERY, SUBSCRIPTION)
 import           Data.Morpheus.Types.Internal.Stream        (Channel (..), Event (..), ResponseEvent (..),
                                                              ResponseStream, StreamChannel, StreamState (..),
@@ -52,26 +50,17 @@ withObject :: ( SelectionSet -> ResolvingStrategy o m e value) -> (Key,ValidSele
 withObject f (_, Selection {selectionRec = SelectionSet selection}) = f selection
 withObject _ (key, Selection {selectionPosition}) = Fail $ subfieldsNotSelected key "" selectionPosition
 
-class Monad m =>
-      GQLFail (t :: (* -> *) -> * -> *) m
-  where
-  gqlFail :: Monad m => Message -> t m a
-  toSuccess :: Monad m => (Message -> b) -> (a -> b) -> t m a -> t m b
+liftM :: (PureOperation o, Monad m) => m a -> Resolver o m e a
+liftM = liftEither . fmap pure
 
-instance Monad m => GQLFail (ExceptT String) m where
-  gqlFail = ExceptT . pure . Left . unpack
-  toSuccess fFail fSuc (ExceptT value) = ExceptT $ pure . mapCases <$> value
-    where
-      mapCases (Right x) = fSuc x
-      mapCases (Left x)  = fFail $ pack $ show x
-
+liftEitherM :: (PureOperation o, Monad m) => m (Either String a) -> Resolver o m e a
+liftEitherM = liftEither
 ----------------------------------------------------------------------------------------
 type ResolveT = ExceptT GQLErrors
 type ResponseT m e  = ResolveT (ResponseStream m e)
 
 --
 -- Recursive Resolver
-
 newtype RecResolver m a b = RecResolver {
   unRecResolver :: a -> ResolveT m b
 }
@@ -126,6 +115,7 @@ instance (PureOperation o, Monad m) => Applicative (ResolvingStrategy o m e) whe
                        pure (f1 <*> res1)
 
 -- GADTResolver
+---------------------------------------------------------------
 data Resolver (o::OperationType) (m :: * -> * ) event value where
     FailedResolver :: { unFailedResolver :: String } -> Resolver o m event value
     QueryResolver:: { unQueryResolver :: ExceptT String m value } -> Resolver QUERY m  event value
@@ -149,7 +139,7 @@ instance Functor m => Functor (Resolver o m e) where
 
 -- GADTResolver Applicative
 instance (PureOperation o ,Monad m) => Applicative (Resolver o m e) where
-    pure = pureRes
+    pure = liftEither . pure . pure
     -------------------------------------
     _ <*> (FailedResolver mErrors) = FailedResolver mErrors
     (FailedResolver mErrors) <*> _ = FailedResolver mErrors
@@ -170,22 +160,22 @@ instance (Monad m) => Monad (Resolver QUERY m e) where
 
 -- Pure Operation
 class PureOperation (o::OperationType) where
-    pureRes :: Monad m => a -> Resolver o m event a
+    liftEither :: Monad m => m (Either String a) -> Resolver o m event a
     pureGraphQLT :: Monad m => a -> ResolvingStrategy o m event a
     eitherGraphQLT :: Monad m => Validation a -> ResolvingStrategy o m event a
 
 instance PureOperation QUERY where
-   pureRes = QueryResolver . pure
+   liftEither = QueryResolver . ExceptT
    pureGraphQLT = QueryResolving . pure
    eitherGraphQLT = QueryResolving . ExceptT . pure
 
 instance PureOperation MUTATION where
-   pureRes = MutResolver [] . pure
+   liftEither = MutResolver [] . ExceptT
    pureGraphQLT = MutationResolving . pure
    eitherGraphQLT = MutationResolving . ExceptT . pure
 
 instance PureOperation SUBSCRIPTION where
-   pureRes = SubResolver []  . const . pure
+   liftEither = SubResolver [] . const . liftEither
    pureGraphQLT = SubscriptionResolving . pure . pure
    eitherGraphQLT = SubscriptionResolving . fmap pure  . ExceptT . pure
 
@@ -257,10 +247,6 @@ toResponseRes (SubscriptionResolving resT)  =
 type family UnSubResolver (a :: * -> *) :: (* -> *)
 
 type instance UnSubResolver (Resolver SUBSCRIPTION m e) = Resolver QUERY m e
-
--------------------------------------------------------------------
-failResolveT :: Monad m => GQLErrors -> ResolveT m a
-failResolveT = ExceptT . pure . Left
 
 -------------------------------------------------------------------
 -- | GraphQL Root resolver, also the interpreter generates a GQL schema from it.